* More portable socket- and DNS lookup handling
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@1194 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/lib/video_libmpeg2.c
@@ -396,9 +396,9 @@ static int init_mpeg2(bgav_stream_t*s)
   priv->frame->strides[2] = priv->info->sequence->chroma_width;
   
   s->description =
-    bgav_sprintf("MPEG-%d, %.2f kb/sec", 
+    bgav_sprintf("MPEG-%d, %d kb/sec", 
                  (priv->info->sequence->flags & SEQ_FLAG_MPEG2) ? 2 : 1,
-                 (float)priv->info->sequence->byte_rate * 8.0 / 1000.0);
+                 (priv->info->sequence->byte_rate * 8) / 1000);
   s->codec_bitrate = priv->info->sequence->byte_rate * 8;
 
   if(!s->timescale)
